backend/requirements.txt
Flask
Flask-JWT-Extended
numpy
pandas
statsmodels
textblob
spacy
tensorflow
scikit-learn
requests
smtplib
twilio
openai
open3d
librosa
paho-mqtt
pytest
face_recognition
web3
qiskit
fastapi
uvicorn